+++ ios/Source/Model/Domain/ApplePayFinalizePayment.swift
@@ -1,0 +1,10 @@
+import Tpay
+
+struct ApplePayFinalizePayment {
+
+    // MARK: - Properties
+
+    let transactionId: String
+    let applePayToken: String
+    let paymentChannel: Headless.Models.PaymentChannel
+}