+++ ios/Source/Model/Transportation/Transportation+ApplePayFinalizePayment.swift
@@ -1,0 +1,10 @@
+extension Transportation {
+
+    struct ApplePayFinalizePayment: Decodable {
+
+        // MARK: - Properties
+
+        let transactionId: String
+        let applePayToken: String
+    }
+}