+++ ios/Source/Model/Transportation/Transportation+ApplePayInitPayment.swift
@@ -1,0 +1,11 @@
+extension Transportation {
+
+    struct ApplePayInitPayment: Decodable {
+
+        // MARK: - Properties
+
+        let paymentDetails: PaymentDetails
+        let payer: T.Payer
+        let callbacks: Callbacks
+    }
+}